Onboard PostgreSQL Flex force delete (#194)
* update create examples to existing flavor

* add the --force flag to delete command

* generate docs

* add forcedeleteinstance waiter

* extract deletion logic to function, test it. Address PR comments

* fix linting, generate docs

* rename util function

* update sdk version

+++ internal/cmd/postgresflex/instance/delete/delete.go
@@ -8,6 +8,7 @@ import (
 	"github.com/stackitcloud/stackit-cli/internal/pkg/confirm"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/errors"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/examples"
+	"github.com/stackitcloud/stackit-cli/internal/pkg/flags"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/globalflags"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/services/postgresflex/client"
 	postgresflexUtils "github.com/stackitcloud/stackit-cli/internal/pkg/services/postgresflex/utils"
@@ -21,23 +22,33 @@ import (
 
 const (
 	instanceIdArg = "INSTANCE_ID"
+
+	forceDeleteFlag = "force"
 )
 
 type inputModel struct {
 	*globalflags.GlobalFlagModel
-	InstanceId string
+	InstanceId  string
+	ForceDelete bool
 }
 
 func NewCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   fmt.Sprintf("delete %s", instanceIdArg),
 		Short: "Deletes a PostgreSQL Flex instance",
-		Long:  "Deletes a PostgreSQL Flex instance.",
-		Args:  args.SingleArg(instanceIdArg, utils.ValidateUUID),
+		Long: fmt.Sprintf("%s\n%s\n%s",
+			"Deletes a PostgreSQL Flex instance.",
+			"By default, instances will be kept in a delayed deleted state for 7 days before being permanently deleted.",
+			"Use the --force flag to force the immediate deletion of a delayed deleted instance.",
+		),
+		Args: args.SingleArg(instanceIdArg, utils.ValidateUUID),
 		Example: examples.Build(
 			examples.NewExample(
 				`Delete a PostgreSQL Flex instance with ID "xxx"`,
 				"$ stackit postgresflex instance delete xxx"),
+			examples.NewExample(
+				`Force the deletion of a delayed deleted PostgreSQL Flex instance with ID "xxx"`,
+				"$ stackit postgresflex instance delete xxx --force"),
 		),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			ctx := context.Background()
@@ -65,33 +76,72 @@ func NewCmd() *cobra.Command {
 				}
 			}
 
-			// Call API
-			req := buildRequest(ctx, model, apiClient)
-			err = req.Execute()
+			toDelete, toForceDelete, err := getNextOperations(ctx, model, apiClient)
 			if err != nil {
-				return fmt.Errorf("delete PostgreSQL Flex instance: %w", err)
+				return err
 			}
 
-			// Wait for async operation, if async mode not enabled
-			if !model.Async {
-				s := spinner.New(cmd)
-				s.Start("Deleting instance")
-				_, err = wait.DeleteInstanceWaitHandler(ctx, apiClient, model.ProjectId, model.InstanceId).WaitWithContext(ctx)
+			if toDelete {
+				// Call API
+				delReq := buildDeleteRequest(ctx, model, apiClient)
+				err = delReq.Execute()
 				if err != nil {
-					return fmt.Errorf("wait for PostgreSQL Flex instance deletion: %w", err)
+					return fmt.Errorf("delete PostgreSQL Flex instance: %w", err)
 				}
-				s.Stop()
+
+				// Wait for async operation, if async mode not enabled
+				if !model.Async {
+					s := spinner.New(cmd)
+					s.Start("Deleting instance")
+					_, err = wait.DeleteInstanceWaitHandler(ctx, apiClient, model.ProjectId, model.InstanceId).WaitWithContext(ctx)
+					if err != nil {
+						return fmt.Errorf("wait for PostgreSQL Flex instance deletion: %w", err)
+					}
+					s.Stop()
+				}
+			}
+
+			if toForceDelete {
+				// Call API
+				forceDelReq := buildForceDeleteRequest(ctx, model, apiClient)
+				err = forceDelReq.Execute()
+				if err != nil {
+					return fmt.Errorf("force delete PostgreSQL Flex instance: %w", err)
+				}
+
+				// Wait for async operation, if async mode not enabled
+				if !model.Async {
+					s := spinner.New(cmd)
+					s.Start("Forcing deletion of instance")
+					_, err = wait.ForceDeleteInstanceWaitHandler(ctx, apiClient, model.ProjectId, model.InstanceId).WaitWithContext(ctx)
+					if err != nil {
+						return fmt.Errorf("wait for PostgreSQL Flex instance force deletion: %w", err)
+					}
+					s.Stop()
+				}
 			}
 
 			operationState := "Deleted"
+			if toForceDelete {
+				operationState = "Forcefully deleted"
+			}
 			if model.Async {
 				operationState = "Triggered deletion of"
+				if toForceDelete {
+					operationState = "Triggered forced deletion of"
+				}
 			}
+
 			cmd.Printf("%s instance %q\n", operationState, instanceLabel)
 			return nil
 		},
 	}
+	configureFlags(cmd)
 	return cmd
+}
+
+func configureFlags(cmd *cobra.Command) {
+	cmd.Flags().BoolP(forceDeleteFlag, "f", false, "Force deletion of a delayed deleted instance")
 }
 
 func parseInput(cmd *cobra.Command, inputArgs []string) (*inputModel, error) {
@@ -105,10 +155,39 @@ func parseInput(cmd *cobra.Command, inputArgs []string) (*inputModel, error) {
 	return &inputModel{
 		GlobalFlagModel: globalFlags,
 		InstanceId:      instanceId,
+		ForceDelete:     flags.FlagToBoolValue(cmd, forceDeleteFlag),
 	}, nil
 }
 
-func buildRequest(ctx context.Context, model *inputModel, apiClient *postgresflex.APIClient) postgresflex.ApiDeleteInstanceRequest {
+func buildDeleteRequest(ctx context.Context, model *inputModel, apiClient *postgresflex.APIClient) postgresflex.ApiDeleteInstanceRequest {
 	req := apiClient.DeleteInstance(ctx, model.ProjectId, model.InstanceId)
 	return req
+}
+
+func buildForceDeleteRequest(ctx context.Context, model *inputModel, apiClient *postgresflex.APIClient) postgresflex.ApiForceDeleteInstanceRequest {
+	req := apiClient.ForceDeleteInstance(ctx, model.ProjectId, model.InstanceId)
+	return req
+}
+
+type PostgreSQLFlexClient interface {
+	GetInstanceExecute(ctx context.Context, projectId, instanceId string) (*postgresflex.InstanceResponse, error)
+	ListVersionsExecute(ctx context.Context, projectId string) (*postgresflex.ListVersionsResponse, error)
+	GetUserExecute(ctx context.Context, projectId, instanceId, userId string) (*postgresflex.GetUserResponse, error)
+}
+
+func getNextOperations(ctx context.Context, model *inputModel, apiClient PostgreSQLFlexClient) (toDelete, toForceDelete bool, err error) {
+	instanceStatus, err := postgresflexUtils.GetInstanceStatus(ctx, apiClient, model.ProjectId, model.InstanceId)
+	if err != nil {
+		return false, false, fmt.Errorf("get PostgreSQL Flex instance status: %w", err)
+	}
+
+	if instanceStatus == wait.InstanceStateDeleted {
+		if !model.ForceDelete {
+			return false, false, fmt.Errorf("instance is already deleted, use --force to force the deletion of a delayed deleted instance")
+		}
+
+		return false, model.ForceDelete, nil
+	}
+
+	return true, model.ForceDelete, nil
 }
